+++ integration/jackson/src/main/java/com/blazebit/persistence/integration/jackson/EntityViewAwareObjectMapper.java
@@ -24,6 +24,7 @@ import com.fasterxml.jackson.databind.BeanDescription;
 import com.fasterxml.jackson.databind.DeserializationConfig;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.deser.BeanDeserializerModifier;
@@ -33,6 +34,7 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 
 import java.lang.reflect.Method;
 import java.util.Collection;
+import java.util.Map;
 
 /**
  * @author Christian Beikov
@@ -61,6 +63,10 @@ public class EntityViewAwareObjectMapper {
             }
         });
         objectMapper.registerModule(module);
+        // We need this property, otherwise Jackson thinks it can use non-visible setters as mutators
+        objectMapper.configure(MapperFeature.INFER_PROPERTY_MUTATORS, false);
+        // Obviously, we don't want fields to be set which are final because these are read-only
+        objectMapper.configure(MapperFeature.ALLOW_FINAL_FIELDS_AS_MUTATORS, false);
         objectMapper.setVisibility(new VisibilityChecker.Std(JsonAutoDetect.Visibility.DEFAULT) {
             @Override
             public boolean isGetterVisible(Method m) {
@@ -80,6 +86,20 @@ public class EntityViewAwareObjectMapper {
             @Override
             public boolean isIsGetterVisible(AnnotatedMethod m) {
                 return metamodel.managedView(m.getDeclaringClass()) == null && super.isGetterVisible(m);
+            }
+
+            // We make setters for collections "invisible" so that is uses a SetterlessProperty to always merge values into existing collections
+
+            @Override
+            public boolean isSetterVisible(Method m) {
+                Class<?> rawParameterType;
+                return super.isSetterVisible(m) && !Collection.class.isAssignableFrom(rawParameterType = m.getParameterTypes()[0]) && !Map.class.isAssignableFrom(rawParameterType);
+            }
+
+            @Override
+            public boolean isSetterVisible(AnnotatedMethod m) {
+                Class<?> rawParameterType;
+                return super.isSetterVisible(m) && !Collection.class.isAssignableFrom(rawParameterType = m.getRawParameterType(0)) && !Map.class.isAssignableFrom(rawParameterType);
             }
         });